Crate rustyhub

rustyhub

Rustyhub is a client for the Github v3 web API, the intent is to create a client that is as complete as possible. Once complete, it will offers an interface to the public and enterprise API. It will also, once complete, offer the capability for using all supported authentication methods, media-types, etc.

Usage

The crate is used by creating a Client and then calling the function associated to the desired endpoint with appropriate arguments. Note: Each function borrows a mutable Client

Example

extern crate rustyhub;

use rustyhub::activity::events;
use rustyhub::auth::auth::Auth;
use rustyhub::client::Client;

fn main() {

    let github_token = "0000000000000000000000000000000000000000".to_string();
    let mut client = Client::new("rusyhub-UserAgent", Auth::OAuth2Token(github_token)));

    let events = client.get_events(&mut client).unwrap();

}

Modules

activity

auth

client

common

error

miscellaneous

utils