Trait rustyhub::activity::events::EventsExt

pub trait EventsExt {
    fn get_events(&mut self) -> Result<Vec<Event>, Error>;
    fn get_repo_events(&mut self,
                       owner: String,
                       repo: String)
                       -> Result<Vec<Event>, Error>;
    fn get_networks_owner_repo_events(&mut self,
                                      owner: String,
                                      repo: String)
                                      -> Result<Vec<Event>, Error>;
    fn get_orgs_org_events(&mut self, org: String) -> Result<Vec<Event>, Error>;
    fn get_users_username_received_events(&mut self,
                                          username: String)
                                          -> Result<Vec<Event>, Error>;
    fn get_users_username_received_events_public(&mut self,
                                                 username: String)
                                                 -> Result<Vec<Event>, Error>;
    fn get_users_username_events(&mut self,
                                 username: String)
                                 -> Result<Vec<Event>, Error>;
    fn get_users_username_events_public(&mut self,
                                        username: String)
                                        -> Result<Vec<Event>, Error>;
    fn get_users_username_events_orgs_org(&mut self,
                                          username: String,
                                          org: String)
                                          -> Result<Vec<Event>, Error>;
}

Required Methods

fn get_events(&mut self) -> Result<Vec<Event>, Error>

[Reference] Returns the list of public events.

Endpoint:

GET /events

fn get_repo_events(&mut self,
owner: String,
repo: String)
-> Result<Vec<Event>, Error>

[Reference] Returns the list of repository events.

Endpoint:

GET /repos/:owner/:repo/events

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn get_networks_owner_repo_events(&mut self,
owner: String,
repo: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list public events for a network of repositories.

Endpoint:

GET /networks/:owner/:repo/events

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn get_orgs_org_events(&mut self, org: String) -> Result<Vec<Event>, Error>

[Reference] Returns a list public events for an organization.

Endpoint:

GET /orgs/:org/events

Parameters:

org: Name of the organization

fn get_users_username_received_events(&mut self,
username: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list events that a user has received.

Endpoint:

GET /users/:username/received_events

Parameters:

username: Name of the user

fn get_users_username_received_events_public(&mut self,
username: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list public events that a user has received.

Endpoint:

GET /users/:username/received_events/public

Parameters:

username: Name of the user

fn get_users_username_events(&mut self,
username: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list events performed by a user.

Endpoint:

GET /users/:username/events

Parameters:

username: Name of the user

fn get_users_username_events_public(&mut self,
username: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list public events performed by a user.

Endpoint:

GET /users/:username/events/public

Parameters:

username: Name of the user

fn get_users_username_events_orgs_org(&mut self,
username: String,
org: String)
-> Result<Vec<Event>, Error>

[Reference] Returns a list events for an organization.

Endpoint:

GET /users/:username/events/orgs/:org

Parameters:

username: Name of the user

org: Name of the organization