Trait rustyhub::activity::notifications::NotificationsExt

pub trait NotificationsExt {
    fn get_notifications(&mut self,
                         all: Option<bool>,
                         participating: Option<bool>,
                         since: Option<String>,
                         before: Option<String>)
                         -> Result<Vec<Notification>, Error>;
    fn get_repos_owner_repo_notifications(&mut self,
                                          owner: String,
                                          repo: String,
                                          all: Option<bool>,
                                          participating: Option<bool>,
                                          since: Option<String>,
                                          before: Option<String>)
                                          -> Result<Vec<Notification>, Error>;
    fn put_notifications(&mut self, last_read_at: String) -> Result<(), Error>;
    fn put_repos_owner_repo_notifications(&mut self,
                                          owner: String,
                                          repo: String,
                                          last_read_at: String)
                                          -> Result<(), Error>;
    fn get_notifications_threads_id(&mut self,
                                    id: String)
                                    -> Result<Notification, Error>;
    fn patch_botifications_threads_id(&mut self,
                                      id: String)
                                      -> Result<(), Error>;
    fn get_notifications_threads_id_subscription(&mut self,
                                                 id: String)
                                                 -> Result<Subscription, Error>;
    fn put_notifications_threads_id_subscription(&mut self,
                                                 id: String,
                                                 subscribed: bool,
                                                 ignored: bool)
                                                 -> Result<Subscription, Error>;
    fn delete_notifications_threads_id_subscription(&mut self,
                                                    id: String)
                                                    -> Result<(), Error>;
}

Required Methods

fn get_notifications(&mut self,
all: Option<bool>,
participating: Option<bool>,
since: Option<String>,
before: Option<String>)
-> Result<Vec<Notification>, Error>

[Reference] Returns the list of notifications.

Endpoint:

GET /notifications

Parameters:

all: Default: false, Show notification marked as read.

participating: Default: false, Only show notifications where the user is directly participating or mentioned.

since: Name of the repository.

before: Name of the repository.

fn get_repos_owner_repo_notifications(&mut self,
owner: String,
repo: String,
all: Option<bool>,
participating: Option<bool>,
since: Option<String>,
before: Option<String>)
-> Result<Vec<Notification>, Error>

[Reference] Returns the list of notifications in a repository.

Endpoint:

GET /notifications

Parameters:

all: Owner of the repository.

participating: Name of the repository.

since: Name of the repository.

before: Name of the repository.

fn put_notifications(&mut self, last_read_at: String) -> Result<(), Error>

[Reference] Marks notifications as read.

Endpoint:

PUT /notifications

Parameters:

last_read_at: Default: "Time.now", The last point at which notifications were checked in the ISO 8601 (YYYY-MM-DDTHH:MM:SSZ) format.

fn put_repos_owner_repo_notifications(&mut self,
owner: String,
repo: String,
last_read_at: String)
-> Result<(), Error>

[Reference] Marks notifications in a repository as read.

Endpoint:

PUT /repos/:owner/:repo/notifications

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

last_read_at: Default: "Time.now", The last point at which notifications were checked in the ISO 8601 (YYYY-MM-DDTHH:MM:SSZ) format.

fn get_notifications_threads_id(&mut self,
id: String)
-> Result<Notification, Error>

[Reference] Returns a single thread.

Endpoint:

GET /notifications/threads/:id

Parameters:

id: Thread id.

fn patch_botifications_threads_id(&mut self, id: String) -> Result<(), Error>

[Reference] Marks a single as thread.

Endpoint:

PATCH /notifications/threads/:id

Parameters:

id: Thread id.

fn get_notifications_threads_id_subscription(&mut self,
id: String)
-> Result<Subscription, Error>

[Reference] Returns a single as thread subscription.

Endpoint:

GET /notifications/threads/:id/subscription

Parameters:

id: Thread id.

fn put_notifications_threads_id_subscription(&mut self,
id: String,
subscribed: bool,
ignored: bool)
-> Result<Subscription, Error>

[Reference] Sets a thread subscription.

Endpoint:

PUT /notifications/threads/:id/subscription

Parameters:

id: Thread id.

subscribed: Determines if notifications should be received from this thread.

ignored: Determines if all notifications should be blocked from this thread.

fn delete_notifications_threads_id_subscription(&mut self,
id: String)
-> Result<(), Error>

[Reference] Deletes a thread subscription. ## Endpoint: DELETE /notifications/threads/:id/subscription ## Parameters: * id: Thread id. Reference: