Trait rustyhub::activity::starring::StarringExt

pub trait StarringExt {
    fn get_repos_owner_repo_stargazers(&mut self,
                                       owner: String,
                                       repo: String)
                                       -> Result<Vec<User>, Error>;
    fn get_repos_owner_repo_stargazers_timestamp(&mut self,
                                                 owner: String,
                                                 repo: String)
                                                 -> Result<Vec<ListStarTimeStamp>, Error>;
    fn get_users_username_starred(&mut self,
                                  username: String,
                                  sort: Option<Sort>,
                                  direction: Option<Direction>)
                                  -> Result<Vec<Repository>, Error>;
    fn get_users_username_starred_timestamp(&mut self,
                                            username: String,
                                            sort: Option<Sort>,
                                            direction: Option<Direction>)
                                            -> Result<Vec<ListRepoStarTimeStamp>, Error>;
    fn get_user_starred(&mut self,
                        sort: Option<Sort>,
                        direction: Option<Direction>)
                        -> Result<Vec<Repository>, Error>;
    fn get_user_starred_timestamp(&mut self,
                                  sort: Option<Sort>,
                                  direction: Option<Direction>)
                                  -> Result<Vec<ListRepoStarTimeStamp>, Error>;
    fn get_user_starred_owner_repo(&mut self,
                                   owner: String,
                                   repo: String)
                                   -> Result<(), Error>;
    fn put_user_starred_owner_repo(&mut self,
                                   owner: String,
                                   repo: String)
                                   -> Result<(), Error>;
    fn delete_user_starred_owner_repo(&mut self,
                                      owner: String,
                                      repo: String)
                                      -> Result<(), Error>;
}

Required Methods

fn get_repos_owner_repo_stargazers(&mut self,
owner: String,
repo: String)
-> Result<Vec<User>, Error>

[Reference] Returns the list of stargazers.

Endpoint:

GET /repos/:owner/:repo/stargazers

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn get_repos_owner_repo_stargazers_timestamp(&mut self,
owner: String,
repo: String)
-> Result<Vec<ListStarTimeStamp>, Error>

[Reference] Returns the list of stargazers with timestamp.

Endpoint:

GET /repos/:owner/:repo/stargazers

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn get_users_username_starred(&mut self,
username: String,
sort: Option<Sort>,
direction: Option<Direction>)
-> Result<Vec<Repository>, Error>

[Reference] Returns the list of repositories being starred by a user.

Endpoint:

GET /users/:username/starred

Parameters:

username: Name of the user.

sort: Default: Sort::Created, sort by when the repository was starred (Sort::Created) or by when the repository was last pushed to (Sort::Updated).

direction: Default: Direction::Ascending, sort in ascending (Direction::Ascending) or in descending (Direction::Descending) order.

fn get_users_username_starred_timestamp(&mut self,
username: String,
sort: Option<Sort>,
direction: Option<Direction>)
-> Result<Vec<ListRepoStarTimeStamp>, Error>

[Reference] Returns the list of repositories with timestamp being starred by a user.

Endpoint:

GET /users/:username/starred

Parameters:

username: Name of the user.

sort: Default: Sort::Created, sort by when the repository was starred (Sort::Created) or by when the repository was last pushed to (Sort::Updated).

direction: Default: Direction::Ascending, sort in ascending (Direction::Ascending) or in descending (Direction::Descending) order.

fn get_user_starred(&mut self,
sort: Option<Sort>,
direction: Option<Direction>)
-> Result<Vec<Repository>, Error>

[Reference] Returns the list of repositories being starred by an authenticated user.

Endpoint:

GET /user/starred

Parameters:

sort: Default: Sort::Created, sort by when the repository was starred (Sort::Created) or by when the repository was last pushed to (Sort::Updated).

direction: Default: Direction::Ascending, sort in ascending (Direction::Ascending) or in descending (Direction::Descending) order.

fn get_user_starred_timestamp(&mut self,
sort: Option<Sort>,
direction: Option<Direction>)
-> Result<Vec<ListRepoStarTimeStamp>, Error>

[Reference] Returns the list of repositories with timestamp being starred by an authenticated user.

Endpoint:

GET /user/starred

Parameters:

sort: Default: Sort::Created, sort by when the repository was starred (Sort::Created) or by when the repository was last pushed to (Sort::Updated).

direction: Default: Direction::Ascending, sort in ascending (Direction::Ascending) or in descending (Direction::Descending) order.

fn get_user_starred_owner_repo(&mut self,
owner: String,
repo: String)
-> Result<(), Error>

[Reference] Returns whether an authenticated user is starring a repository.

Endpoint:

GET /user/starred/:owner/:repo

Parameters:

owner: Owner of the repository.

repo: Name of the repository. ## Return Values:

If repository is starred: returns Ok(())

If repository is not starred: returns Error::Github

fn put_user_starred_owner_repo(&mut self,
owner: String,
repo: String)
-> Result<(), Error>

[Reference] Stars a repository.

Endpoint:

PUT /user/starred/:owner/:repo

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn delete_user_starred_owner_repo(&mut self,
owner: String,
repo: String)
-> Result<(), Error>

[Reference] Unstars a repository.

Endpoint:

DELETE /user/starred/:owner/:repo

Parameters:

owner: Owner of the repository.

repo: Name of the repository.