Trait rustyhub::activity::watching::WatchingExt

pub trait WatchingExt {
    fn get_repos_owner_repo_subscribers(&mut self,
                                        owner: String,
                                        repo: String)
                                        -> Result<User, Error>;
    fn get_users_username_subscription(&mut self,
                                       username: String)
                                       -> Result<Repository, Error>;
    fn get_user_subscription(&mut self) -> Result<Repository, Error>;
    fn get_repos_owner_repo_subscription(&mut self,
                                         owner: String,
                                         repo: String)
                                         -> Result<Subscription, Error>;
    fn put_repos_owner_repo_subscription(&mut self,
                                         owner: String,
                                         repo: String,
                                         subscribed: bool,
                                         ignored: bool)
                                         -> Result<Subscription, Error>;
    fn delete_repos_owner_subscription(&mut self,
                                       owner: String,
                                       repo: String)
                                       -> Result<(), Error>;
}

Required Methods

fn get_repos_owner_repo_subscribers(&mut self,
owner: String,
repo: String)
-> Result<User, Error>

[Reference] Returns the list of watchers.

Endpoint:

GET /repos/:owner/:repo/subscribers

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn get_users_username_subscription(&mut self,
username: String)
-> Result<Repository, Error>

[Reference] Returns the list of repositories being watched by a user.

Endpoint:

GET /users/:username/subscriptions

Parameters:

username: Name of the user.

fn get_user_subscription(&mut self) -> Result<Repository, Error>

[Reference] Returns the list of repositories being watched by an authenticated user.

Endpoint:

GET /user/subscriptions

fn get_repos_owner_repo_subscription(&mut self,
owner: String,
repo: String)
-> Result<Subscription, Error>

[Reference] Returns the repository subscription.

Endpoint:

GET /repos/:owner/:repo/subscription

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

fn put_repos_owner_repo_subscription(&mut self,
owner: String,
repo: String,
subscribed: bool,
ignored: bool)
-> Result<Subscription, Error>

[Reference] Sets a repository subscription.

Endpoint:

PUT /repos/:owner/:repo/subscription

Parameters:

owner: Owner of the repository.

repo: Name of the repository.

subscribed: Determines if notifications should be received from this repository.

ignored: Determines if all notifications should be blocked from this repository.

fn delete_repos_owner_subscription(&mut self,
owner: String,
repo: String)
-> Result<(), Error>

[Reference] Deletes a repository subscription/ unfollows a repository.

Endpoint:

DELETE /repos/:owner/:repo/subscription

Parameters:

owner: Owner of the repository.

repo: Name of the repository.