Struct rustyhub::client::Client

pub struct Client {
    pub http_client: HyperClient,
    pub api_url: String,
    pub user_agent: String,
    pub authentication: Auth,
}

Struct with client state, values in this struct may be modified at any time, the new values will be used from the next API call Handles averything related to HTTP

Fields
http_client: HyperClient

Internal http client.

api_url: String

Base URL to the API, can be modified to use the enterprise API.

user_agent: String

Value used for the User-Agent key in request headers.

authentication: Auth

Optional authorization token, acquired from https://github.com/settings/tokens, should be set to None when purely accessing endpoints requiring no authentication.

Methods

impl Client

fn new(user_agent: &str, auth: Auth) -> Client

Creates a Client state with the default API URL.

fn with_url(url: &str, user_agent: &str, auth: Auth) -> Client

Creates a Client state an API URL other than the default.

fn get_default_headers(&self) -> Headers

Returns the default headers used throughout the crate.

fn response_to_string(response: &mut Response) -> Result<String, Error>

Utility to extract the body of an API response.

fn get(&self,
endpoint: String,
headers: Option<Headers>)
-> Result<Response, Error>

GET request with optionally overridden headers.

fn post(&self,
endpoint: String,
headers: Option<Headers>)
-> Result<Response, Error>

POST request with optionally overridden headers.

fn put(&self,
endpoint: String,
headers: Option<Headers>)
-> Result<Response, Error>

PUT request with optionally overridden headers.

fn patch(&self,
endpoint: String,
headers: Option<Headers>)
-> Result<Response, Error>

PATCH request with optionally overridden headers.

fn delete(&self,
endpoint: String,
headers: Option<Headers>)
-> Result<Response, Error>

DELETE request with optionally overridden headers.

fn get_body(&self,
endpoint: String,
headers: Option<Headers>,
body: String)
-> Result<Response, Error>

GET request with optionally overridden headers and a body.

fn post_body(&self,
endpoint: String,
headers: Option<Headers>,
body: String)
-> Result<Response, Error>

POST request with optionally overridden headers and a body.

fn put_body(&self,
endpoint: String,
headers: Option<Headers>,
body: String)
-> Result<Response, Error>

PUT request with optionally overridden headers and a body.

fn patch_body(&self,
endpoint: String,
headers: Option<Headers>,
body: String)
-> Result<Response, Error>

PATCH request with optionally overridden headers and a body.

fn delete_body(&self,
endpoint: String,
headers: Option<Headers>,
body: String)
-> Result<Response, Error>

DELETE request with optionally overridden headers and a body.